+++ CHANGELOG.md
@@ -1,2 +1,4 @@
-0.21.0 - 
+0.21.0 - CrazyLogin Migration
+
+This release adds migration from CrazyLogin.